rename ConnectorResourceService to StorgeResourceService
