+++ eventmesh-storage/eventmesh-storage-rabbitmq/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.ConnectorResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-rabbitmq=org.apache.eventmesh.storage.rabbitmq.connector.RabbitmqConnectorResourceService
+rabbitmq=org.apache.eventmesh.storage.rabbitmq.connector.RabbitmqStorageResourceService
