+++ eventmesh-storage/eventmesh-storage-redis/src/main/java/org/apache/eventmesh/storage/redis/connector/RedisStorageResourceService.java
@@ -8,18 +8,18 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  */
 
-package org.apache.eventmesh.storage.standalone.resource;
+package org.apache.eventmesh.storage.redis.connector;
 
-import org.apache.eventmesh.api.storage.ConnectorResourceService;
+import org.apache.eventmesh.api.storage.StorageResourceService;
 
-public class StandaloneConnectorResourceService implements ConnectorResourceService {
+public class RedisStorageResourceService implements StorageResourceService {
 
     @Override
     public void init() throws Exception {
