+++ eventmesh-storage/eventmesh-storage-redis/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.ConnectorResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-redis=org.apache.eventmesh.storage.redis.connector.RedisConnectorResourceService
+redis=org.apache.eventmesh.storage.redis.connector.RedisStorageResourceService
