+++ eventmesh-storage/eventmesh-storage-rocketmq/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.ConnectorResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-rocketmq=org.apache.eventmesh.storage.rocketmq.connector.ConnectorResourceServiceRocketmqImpl
+rocketmq=org.apache.eventmesh.storage.rocketmq.connector.StorageResourceServiceRocketmqImpl
