+++ eventmesh-storage/eventmesh-storage-standalone/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.ConnectorResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-standalone=org.apache.eventmesh.storage.standalone.resource.StandaloneConnectorResourceService
+standalone=org.apache.eventmesh.storage.standalone.resource.StandaloneStorageResourceService
